requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-timeout==3.0.1
attrs==21.4.0
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.5
cryptography==37.0.2
datadispatch==1.0.0
discord==1.7.3
discord-components==2.1.2
discord-py-slash-command==3.0.3
discord.ext==0.7
discord.py==1.7.3
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
paramiko==2.11.0
pycparser==2.21
PyNaCl==1.5.0
remoteauthclient==1.2
requests==2.28.0
sellix-python-sdk==1.0.1
six==1.16.0
tasks==2.8.0
typing_extensions==4.2.0
urllib3==1.26.9
websockets==10.3
yarl==1.7.2
